kv: move mock interface to test files (#15659)
