+++ kv/fault_injection_test.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kv_test
+package kv
 
 import (
 	"context"
@@ -19,7 +19,6 @@ import (
 	. "github.com/pingcap/check"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/terror"
-	"github.com/pingcap/tidb/kv"
 )
 
 type testFaultInjectionSuite struct{}
@@ -27,17 +26,17 @@ type testFaultInjectionSuite struct{}
 var _ = Suite(testFaultInjectionSuite{})
 
 func (s testFaultInjectionSuite) TestFaultInjectionBasic(c *C) {
-	var cfg kv.InjectionConfig
+	var cfg InjectionConfig
 	err1 := errors.New("foo")
 	cfg.SetGetError(err1)
 	cfg.SetCommitError(err1)
 
-	storage := kv.NewInjectedStore(kv.NewMockStorage(), &cfg)
+	storage := NewInjectedStore(newMockStorage(), &cfg)
 	txn, err := storage.Begin()
 	c.Assert(err, IsNil)
 	_, err = storage.BeginWithStartTS(0)
 	c.Assert(err, IsNil)
-	ver := kv.Version{Ver: 1}
+	ver := Version{Ver: 1}
 	snap, err := storage.GetSnapshot(ver)
 	c.Assert(err, IsNil)
 	b, err := txn.Get(context.TODO(), []byte{'a'})
@@ -61,7 +60,7 @@ func (s testFaultInjectionSuite) TestFaultInjectionBasic(c *C) {
 	cfg.SetGetError(nil)
 	cfg.SetCommitError(nil)
 
-	storage = kv.NewInjectedStore(kv.NewMockStorage(), &cfg)
+	storage = NewInjectedStore(newMockStorage(), &cfg)
 	txn, err = storage.Begin()
 	c.Assert(err, IsNil)
 	snap, err = storage.GetSnapshot(ver)
@@ -76,15 +75,14 @@ func (s testFaultInjectionSuite) TestFaultInjectionBasic(c *C) {
 	c.Assert(bs, IsNil)
 
 	b, err = snap.Get(context.TODO(), []byte{'a'})
-	c.Assert(terror.ErrorEqual(kv.ErrNotExist, err), IsTrue)
+	c.Assert(terror.ErrorEqual(ErrNotExist, err), IsTrue)
 	c.Assert(b, IsNil)
 
-	bs, err = snap.BatchGet(context.Background(), []kv.Key{[]byte("a")})
+	bs, err = snap.BatchGet(context.Background(), []Key{[]byte("a")})
 	c.Assert(err, IsNil)
 	c.Assert(len(bs), Equals, 0)
 
 	err = txn.Commit(context.Background())
 	c.Assert(err, NotNil)
-	c.Assert(terror.ErrorEqual(err, kv.ErrTxnRetryable), IsTrue)
-
+	c.Assert(terror.ErrorEqual(err, ErrTxnRetryable), IsTrue)
 }
