+++ kv/interface_mock_test.go
@@ -117,8 +117,8 @@ func (t *mockTxn) SetVars(vars *Variables) {
 
 }
 
-// NewMockTxn new a mockTxn.
-func NewMockTxn() Transaction {
+// newMockTxn new a mockTxn.
+func newMockTxn() Transaction {
 	return &mockTxn{
 		opts:  make(map[Option]interface{}),
 		valid: true,
@@ -130,7 +130,7 @@ type mockStorage struct {
 }
 
 func (s *mockStorage) Begin() (Transaction, error) {
-	return NewMockTxn(), nil
+	return newMockTxn(), nil
 }
 
 func (*mockTxn) IsPessimistic() bool {
@@ -185,14 +185,8 @@ func (s *mockStorage) ShowStatus(ctx context.Context, key string) (interface{}, 
 	return nil, nil
 }
 
-// MockTxn is used for test cases that need more interfaces than Transaction.
-type MockTxn interface {
-	Transaction
-	GetOption(opt Option) interface{}
-}
-
-// NewMockStorage creates a new mockStorage.
-func NewMockStorage() Storage {
+// newMockStorage creates a new mockStorage.
+func newMockStorage() Storage {
 	return &mockStorage{}
 }
