+++ kv/mock_test.go
@@ -15,6 +15,7 @@ package kv
 
 import (
 	"context"
+
 	. "github.com/pingcap/check"
 )
 
@@ -24,7 +25,7 @@ type testMockSuite struct {
 }
 
 func (s testMockSuite) TestInterface(c *C) {
-	storage := NewMockStorage()
+	storage := newMockStorage()
 	storage.GetClient()
 	storage.UUID()
 	version, err := storage.CurrentVersion()
